update: [이승현] 관리자는 페이지 단위로 회원 조회를 할 수 있습니다 [JSAB-88]
 requestBody로 입력된 값 처리 normalizedLastId의 변환처리가 정상적으로 동작하지않기에 longValue 대입
입력된 값을 치환처리하는 변수 선언에 final 추가

+++ src/main/java/com/wowraid/jobspoon/administer/controller/dto/AdministratorUserInfoRequest.java
@@ -19,7 +19,8 @@ public class AdministratorUserInfoRequest {
     private Long lastAccountId;
 
     private static final Set<Integer> ALLOWED_SIZES = Set.of(30,50,70);
-    public long normalizedLastId() { return lastAccountId == null ? 0L : lastAccountId; }
+    public long normalizedLastId() {
+        return lastAccountId == null ? 0L : lastAccountId.longValue(); }
     public void validate() {
         if(!ALLOWED_SIZES.contains(pageSize)) {
             throw new InvalidParameterException("Invalid pagesize ->size must be one of 30,50,70");
